+++ src/commands/lambda/__tests__/functions/commons.test.ts
@@ -13,7 +13,7 @@ import {mockClient} from 'aws-sdk-client-mock'
 
 import 'aws-sdk-client-mock-jest'
 import {CI_API_KEY_ENV_VAR, CI_SITE_ENV_VAR} from '../../../../constants'
-import {MOCK_DATADOG_API_KEY} from '../../../../helpers/__tests__/serverlessFixtures'
+import {MOCK_DATADOG_API_KEY} from '../../../../helpers/__tests__/fixtures'
 import {maskEnvVar} from '../../../../helpers/utils'
 
 import {
@@ -30,6 +30,7 @@ import {
   MERGE_XRAY_TRACES_ENV_VAR,
   Runtime,
   TRACE_ENABLED_ENV_VAR,
+  SKIP_MASKING_LAMBDA_ENV_VARS,
 } from '../../constants'
 import {
   addLayerArn,
