+++ src/commands/lambda/__tests__/instrument.test.ts
@@ -18,12 +18,13 @@ import {mockClient} from 'aws-sdk-client-mock'
 import 'aws-sdk-client-mock-jest'
 import {Cli} from 'clipanion/lib/advanced'
 
-import {AWS_DEFAULT_REGION_ENV_VAR, CI_API_KEY_ENV_VAR, CI_SITE_ENV_VAR} from '../../../constants'
-import {createMockContext, MOCK_DATADOG_API_KEY} from '../../../helpers/__tests__/serverlessFixtures'
+import {CI_API_KEY_ENV_VAR, CI_SITE_ENV_VAR} from '../../../constants'
+import {createMockContext, MOCK_DATADOG_API_KEY} from '../../../helpers/__tests__/fixtures'
 import {requestConfirmation} from '../../../helpers/prompt'
 
 import {
   AWS_ACCESS_KEY_ID_ENV_VAR,
+  AWS_DEFAULT_REGION_ENV_VAR,
   AWS_SECRET_ACCESS_KEY_ENV_VAR,
   AWS_SESSION_TOKEN_ENV_VAR,
   DEFAULT_LAYER_AWS_ACCOUNT,
