+++ src/commands/lambda/__tests__/uninstrument.test.ts
@@ -21,8 +21,8 @@ import {
 import {fromIni} from '@aws-sdk/credential-providers'
 import {mockClient} from 'aws-sdk-client-mock'
 
-import {AWS_DEFAULT_REGION_ENV_VAR, SITE_ENV_VAR} from '../../../constants'
-import {createMockContext} from '../../../helpers/__tests__/serverlessFixtures'
+import {SITE_ENV_VAR} from '../../../constants'
+import {createMockContext} from '../../../helpers/__tests__/fixtures'
 import {requestConfirmation} from '../../../helpers/prompt'
 
 import 'aws-sdk-client-mock-jest'
@@ -30,6 +30,7 @@ import 'aws-sdk-client-mock-jest'
 import {
   APM_FLUSH_DEADLINE_MILLISECONDS_ENV_VAR,
   AWS_ACCESS_KEY_ID_ENV_VAR,
+  AWS_DEFAULT_REGION_ENV_VAR,
   AWS_SECRET_ACCESS_KEY_ENV_VAR,
   ENVIRONMENT_ENV_VAR,
   FLUSH_TO_LOG_ENV_VAR,
