+++ src/commands/lambda/flare.ts
@@ -14,13 +14,13 @@ import {AwsCredentialIdentity} from '@aws-sdk/types'
 import chalk from 'chalk'
 import {Command} from 'clipanion'
 
-import {API_KEY_ENV_VAR, AWS_DEFAULT_REGION_ENV_VAR, CI_API_KEY_ENV_VAR, FLARE_OUTPUT_DIRECTORY} from '../../constants'
-import {createDirectories, deleteFolder, writeFile, zipContents} from '../../helpers/fileSystem'
-import {sendToDatadog} from '../../helpers/flareFunctions'
+import {API_KEY_ENV_VAR, CI_API_KEY_ENV_VAR, FLARE_OUTPUT_DIRECTORY} from '../../constants'
+import {sendToDatadog} from '../../helpers/flare'
+import {createDirectories, deleteFolder, writeFile, zipContents} from '../../helpers/fs'
 import {requestConfirmation} from '../../helpers/prompt'
 import * as helpersRenderer from '../../helpers/renderer'
 
-import {PROJECT_FILES} from './constants'
+import {AWS_DEFAULT_REGION_ENV_VAR, PROJECT_FILES, SKIP_MASKING_LAMBDA_ENV_VARS} from './constants'
 import {getAWSCredentials, getLambdaFunctionConfig, getRegion, maskStringifiedEnvVar} from './functions/commons'
 import {requestAWSCredentials, requestFilePath} from './prompt'
 import * as commonRenderer from './renderers/common-renderer'
