+++ src/commands/lambda/instrument.ts
@@ -4,7 +4,6 @@ import {AwsCredentialIdentity} from '@aws-sdk/types'
 import {bold} from 'chalk'
 import {Cli, Command} from 'clipanion'
 
-import {AWS_DEFAULT_REGION_ENV_VAR} from '../../constants'
 import {requestConfirmation} from '../../helpers/prompt'
 import * as helperRenderer from '../../helpers/renderer'
 import {resolveConfigFromFile, filterAndFormatGithubRemote, DEFAULT_CONFIG_PATHS} from '../../helpers/utils'
@@ -12,7 +11,13 @@ import {resolveConfigFromFile, filterAndFormatGithubRemote, DEFAULT_CONFIG_PATHS
 import {getCommitInfo, newSimpleGit} from '../git-metadata/git'
 import {UploadCommand} from '../git-metadata/upload'
 
-import {ENVIRONMENT_ENV_VAR, EXTRA_TAGS_REG_EXP, SERVICE_ENV_VAR, VERSION_ENV_VAR} from './constants'
+import {
+  AWS_DEFAULT_REGION_ENV_VAR,
+  ENVIRONMENT_ENV_VAR,
+  EXTRA_TAGS_REG_EXP,
+  SERVICE_ENV_VAR,
+  VERSION_ENV_VAR,
+} from './constants'
 import {
   checkRuntimeTypesAreUniform,
   coerceBoolean,
