+++ src/commands/lambda/prompt.ts
@@ -2,12 +2,13 @@ import {blueBright, bold, dim} from 'chalk'
 import {filter} from 'fuzzy'
 import inquirer from 'inquirer'
 
-import {AWS_DEFAULT_REGION_ENV_VAR, CI_API_KEY_ENV_VAR, CI_SITE_ENV_VAR, DATADOG_SITES} from '../../constants'
+import {CI_API_KEY_ENV_VAR, CI_SITE_ENV_VAR, DATADOG_SITES} from '../../constants'
 import {isValidDatadogSite} from '../../helpers/validation'
 
 import {
   AWS_ACCESS_KEY_ID_ENV_VAR,
   AWS_ACCESS_KEY_ID_REG_EXP,
+  AWS_DEFAULT_REGION_ENV_VAR,
   AWS_SECRET_ACCESS_KEY_ENV_VAR,
   AWS_SECRET_ACCESS_KEY_REG_EXP,
   AWS_SECRET_ARN_REG_EXP,
